chore: create a README-VANILLA.md
